--- a/pyproject.toml
+++ b/pyproject.toml
@@ -1,8 +1,8 @@
 [project]
 name = "jaxloudnorm"
 requires-python = "~=3.9"
-version = "0.1.0"
-description = "Implementation of ITU-R BS.1770-4 loudness algorithm in Jax in Python."
+version = "0.2.0"
+description = "Implementation of ITU-R BS.1770-4 loudness algorithm in JAX in Python."
 authors = [
     { name = "Boris Kuznetsov" },
     { email = "boris.kuznetsov@mailbox.org" },