@DBraun 2025-01-29
Thanks! Next question... would you be open to putting this on PyPI? I would like to have some of my other PyPI projects depend on it.